matcher_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.1
chromadb==0.6.3
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.2
python-multipart==0.0.20